added sales app
