+++ cashflow/settings/base.py
@@ -40,6 +40,7 @@ USER_APPS = [
     'users.apps.UsersConfig',
     'branchoffices.apps.BranchofficesConfig',
     'products.apps.ProductsConfig',
+    'sales.apps.SalesConfig',
 ]
 
 INSTALLED_APPS = DJANGO_APPS + USER_APPS + THIRD_APPS
