+++ products/urls.py
@@ -8,11 +8,6 @@ urlpatterns = [
     # test
     url(r'^test/$', views.test, name='test'),
 
-    # sales
-    url(r'^sales/$', views.sales, name='sales'),
-    url(r'^sales/new/$', views.new_sale, name='new_sale'),
-    url(r'^sales/sales-day/$', views.get_sales_day_view, name='get_sales_day'),
-
     # Supplies
     url(r'^supplies/$', views.supplies, name='supplies'),
     url(r'^supplies/new/$', views.new_supply, name='new_supply'),
