+++ sales/forms.py
@@ -1,0 +1,4 @@
+# -*- encoding: utf-8 -*-
+from __future__ import unicode_literals
+from django import forms
+
