+++ sales/urls.py
@@ -1,0 +1,12 @@
+from django.conf.urls import url
+
+from sales import views
+
+app_name = 'sales'
+
+urlpatterns = [
+    # sales
+    url(r'^sales/$', views.sales, name='sales'),
+    url(r'^sales/new/$', views.new_sale, name='new_sale'),
+    url(r'^sales/sales-day/$', views.get_sales_day_view, name='get_sales_day'),
+]
